ml/requirements.txt
redis
google-cloud-pubsub